Move build commands to setuputils
